refact: extracts common functions
